+++ src/components/Card/Card.test.tsx
@@ -66,7 +66,6 @@ describe("Card", () => {
       id="regularCard"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-regularCard",
         src: getPlaceholderImage("smaller", 0),
       }}
       ref={ref}
@@ -88,7 +87,6 @@ describe("Card", () => {
       id="cardWithExtendedStyles"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-cardWithExtendedStyles",
         src: getPlaceholderImage("smaller", 0),
       }}
     >
@@ -125,7 +123,6 @@ describe("Card", () => {
       id="cardWithNoCTAs"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-cardWithNoCTAs",
         src: getPlaceholderImage("smaller", 0),
       }}
     >
@@ -145,7 +142,6 @@ describe("Card", () => {
       id="cardWithNoContent"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-cardWithNoContent",
         src: getPlaceholderImage("smaller", 0),
       }}
     >
@@ -195,7 +191,6 @@ describe("Card", () => {
       id="fullclick"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-fullclick",
         src: getPlaceholderImage("smaller", 0),
       }}
       mainActionLink="http://nypl.org"
@@ -227,10 +222,9 @@ describe("Card", () => {
   );
   const cardWithRightActions = () => (
     <Card
-      id="cardID"
+      id="cardWithRightActions"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-cardWithRightActions",
         src: getPlaceholderImage("smaller", 0),
       }}
       isAlignedRightActions
@@ -256,7 +250,6 @@ describe("Card", () => {
       id="chakraProps"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-chakraProps",
         src: getPlaceholderImage("smaller", 0),
       }}
       p="s"
@@ -278,7 +271,6 @@ describe("Card", () => {
       id="otherProps"
       imageProps={{
         alt: "Alt text",
-        id: "img-id-otherProps",
         src: getPlaceholderImage("smaller", 0),
       }}
       data-testid="card-testid"