+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Changes the names of theme objects that conflict with Chakra's base names from `Custom` to `Reservoir` for consistency. This impacts theme objects for `Breadcrumbs`, `Button`, `Select`, `Slider`, and `Table`.
 - Replaces positional function args with object references in utility functions/components when there are 3 or more arguments.
 - Updates `Table`'s `tableTextSize` prop to accept `caption`.
+- Updates the `id` value for all components to use the `ds-[componentName]-${additionalUserID}` format.
 
 ### Removals